requirements.txt
pytest
pytest-asyncio
pytest-order
allure-pytest
python-dotenv
pyyaml
playwright
